requirements.txt
PyQt5==5.15.9
opencv-python==4.8.1.78
pygame==2.5.2
numpy==1.24.3
Pillow==10.0.1